mcp-scaffolds/trading-mt5-mcp/requirements.txt
mcp>=1.0.0
MetaTrader5>=5.0.45
python-dotenv>=1.0.1
pydantic>=2.6
requests>=2.31.0
pytest>=8.0
